+++ NEWS.md
@@ -1,3 +1,7 @@
+# CoastCR 1.1.0
+
+* Currently version. The stats summary are saving as csv file and include quantiles information.
+
 # CoastCR 1.0.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* Original version. Results summary save as png format file.